streamlink: update to 0.4.0
0.4.0 of Streamlink!

114 commits since the last release and a lot has changed.

In general, we've added some localization as well as an assortment of new plugins.

We've also introduced a change for Streamlink to not check for new updates each time Streamlink starts. We found this feature annoying as well as delaying the initial start of the stream. This feature can be re-enabled by the command line.

The major features of this release are:

    New plugins added
    Ongoing support to current plugins via bug fixes
    Ensure retries to HLS streams
    Disable update check

Many thanks to all contributors who have contributed in this release!