Allow bootstrapping without cwrappers
Since cwrappers was enabled by default on some platforms, it became
effectively impossible to bootstrap without cwrappers there.

With this change "--cwrappers" needs to be set to "auto", "yes" or "no"
explicitly when used.

No functional change intended without specifying "--cwrappers" (or with
"auto").

Submitted on tech-pkg@ with no objection.